Add resize-rootfs recipe
+ add resize-rootfs recipe which simplifies resizing rootfs
  partition on SD card;

The "resize-rootfs" commands needs to be called twice: the first
time it will change the partition table and reboot the system
and the second time it will resize the file system to fill the
partition.  This allows smaller SD card binary images whilst
providing an easy solution to use all the available space.

Signed-off-by: Guillaume Tucker <guillaume.tucker@arm.com>